Couple more methods left to move to the base class
