+++ ClimateExplorer.Visualiser/Pages/RegionalAndGlobal.razor.cs
@@ -20,31 +20,13 @@ public partial class RegionalAndGlobal : ChartablePage
     
     Modal addDataSetModal { get; set; }
 
-    protected override async Task OnInitializedAsync()
-    {
-        if (base.DataService == null)
-        {
-            throw new NullReferenceException(nameof(base.DataService));
-        }
-        DataSetDefinitions = (await base.DataService.GetDataSetDefinitions()).ToList();
-    }
-
-
     public void Dispose()
     {
         //Logger.LogInformation("Instance " + _componentInstanceId + " disposing");
         //   NavManager.LocationChanged -= HandleLocationChanged;
     }
 
-    async Task OnAddDataSet(DataSetLibraryEntry dle)
-    {
-        await base.chartView.OnAddDataSet(dle, DataSetDefinitions);
-    }
 
-    async Task OnChartPresetSelected(List<ChartSeriesDefinition> chartSeriesDefinitions)
-    {
-        await base.chartView.OnChartPresetSelected(chartSeriesDefinitions);
-    }
 
     string GetPageTitle()
     {
@@ -52,7 +34,7 @@ public partial class RegionalAndGlobal : ChartablePage
 
         string title = $"ClimateExplorer";// {locationText}";
 
-        base.Logger.LogInformation("GetPageTitle() returning '" + title + "' NavigateTo");
+        Logger.LogInformation("GetPageTitle() returning '" + title + "' NavigateTo");
 
         return title;
     }
@@ -62,18 +44,7 @@ public partial class RegionalAndGlobal : ChartablePage
         return addDataSetModal.Show();
     }
 
-    protected async Task OnDownloadDataClicked(DataDownloadPackage dataDownloadPackage)
+    protected override async Task UpdateOtherViews()
     {
-        var fileStream = Exporter.ExportChartData(Logger, new List<Location>(), dataDownloadPackage, NavManager.Uri.ToString());
-
-        var locationNames = dataDownloadPackage.ChartSeriesWithData.SelectMany(x => x.ChartSeries.SourceSeriesSpecifications).Select(x => x.LocationName).Where(x => x != null).Distinct().ToArray();
-
-        var fileName = locationNames.Any() ? string.Join("-", locationNames) + "-" : "";
-
-        fileName = $"Export-{fileName}-{dataDownloadPackage.BinGranularity}-{dataDownloadPackage.Bins.First().Label}-{dataDownloadPackage.Bins.Last().Label}.csv";
-
-        using var streamRef = new DotNetStreamReference(stream: fileStream);
-
-        await base.JsRuntime.InvokeVoidAsync("downloadFileFromStream", fileName, streamRef);
     }
 }
